add command line parsing with dry-run

+++ sync.py
@@ -1,12 +1,18 @@
 #!/usr/bin/env python
 
-import shutil, os, sys
+import shutil, os, argparse
 
+parser = argparse.ArgumentParser()
+parser.add_argument('-n', '--dry-run', action='store_true')
+parser.add_argument('-v', '--verbose', action='store_true')
+parser.add_argument('folders', type=str, nargs=2)
+args = parser.parse_args()
 class Sync:
-  def __init__(self, path1, path2):
+  def __init__(self, path1, path2, dry_run=False):
      self.path1 = path1
      self.path2 = path2
      self.notinitialdir = False
+     self.dry_run = dry_run
      self.walk(self.path1)
   def splitAndForm(self, path):
     return os.path.join(self.path2,path.replace(self.path1, '', 1))
@@ -24,34 +30,45 @@ class Sync:
       filename = os.path.abspath(filename)
       file2 = self.splitAndForm(filename)
       if not os.path.exists(file2):
-        print 'Copying', file2
-        shutil.copy2(filename, file2)
+        if self.dry_run:
+          print 'Does not exist', file2
+        else:
+          print filename, '->', file2
+          shutil.copy2(filename, file2)
         continue
       time1 = os.path.getmtime(filename)
       time2 = os.path.getmtime(file2)
       delta = time1 - time2
       if abs(delta) < 2:
-        pass
+        continue
       elif delta > 0:
-#        print time1, '->', time2
-        print filename, '->', file2
-        os.remove(file2)
-        shutil.copy2(filename, file2)
+        if self.dry_run:
+          print 'Contents changed', filename
+        else:
+          print filename, '->', file2
+          os.remove(file2)
+          shutil.copy2(filename, file2)
       else:
-#        print time1, '->', time2
-        print file2, '->', filename
-        os.remove(filename)
-        shutil.copy2(file2, filename)
+        if self.dry_run:
+          print 'Contents changed', file2
+        else:
+          print file2, '->', filename
+          os.remove(filename)
+          shutil.copy2(file2, filename)
    
   def walk(self, folder):
-    print 'In', folder
+    if args.verbose:
+      print 'In', folder
     os.chdir(folder)
     files, dirs = self.splitDirsAndFiles(folder)
     folder2 = self.splitAndForm(folder)
     recurse = True
     if (not os.path.exists(folder2)) and self.notinitialdir:
-      print 'Copying', folder2
-      shutil.copytree(folder, folder2)
+      if self.dry_run:
+        print 'Does not exist', folder2
+      else:
+        print folder, '->', folder2
+        shutil.copytree(folder, folder2)
       recurse = False
     self.notinitialdir = True
     if recurse:
@@ -60,11 +77,4 @@ class Sync:
       os.chdir(folder)
       self.walk(os.path.abspath(directory))
 
-if __name__ == '__main__':
-  if len(sys.argv) > 2:
-    path1 = sys.argv[1]
-    path2 = sys.argv[2]
-  else:
-    path1 = '/home/ateet/Documents/codes/temp/syn1/'
-    path2 = '/home/ateet/Documents/codes/temp/syn2/'
-  sync_obj = Sync(path1, path2)
+Sync(args.folders[0], args.folders[1], args.dry_run)
